backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.4.0
beanie==1.26.0
pydantic==2.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
pdfplumber==0.11.0
python-docx==1.1.2
sentence-transformers==3.0.1
scikit-learn==1.5.0
openpyxl==3.1.4
